@@ -7,58 +7,57 @@ using Trakx.Utils.Testing;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Trakx.Fireblocks.ApiClient.Tests.Integration
+namespace Trakx.Fireblocks.ApiClient.Tests.Integration;
+
+public class ExternalWalletsClientTests : FireblocksClientTestsBase
 {
-    public class ExternalWalletsClientTests : FireblocksClientTestsBase
+    private readonly IExternalWalletsClient _externalWalletsClient;
+    private readonly MockCreator _mockCreator;
+
+    public ExternalWalletsClientTests(FireblocksApiFixture apiFixture, ITestOutputHelper output) 
+        : base(apiFixture, output)
     {
-        private readonly IExternalWalletsClient _externalWalletsClient;
-        private readonly MockCreator _mockCreator;
+        _externalWalletsClient = ServiceProvider.GetRequiredService<IExternalWalletsClient>();
+        _mockCreator = new MockCreator(output);
+    }
 
-        public ExternalWalletsClientTests(FireblocksApiFixture apiFixture, ITestOutputHelper output) 
-            : base(apiFixture, output)
+    [Fact]
+    public async Task GetExternalWalletsAsync_should_return_all_wallets()
+    {
+        var response = await _externalWalletsClient.GetExternalWalletsAsync();
+        response.Result.Should().NotBeEmpty();
+    }
+
+    [Fact]
+    public async Task GetExternalWalletAsync_should_return_one_particular_wallet()
+    {
+        var response = await _externalWalletsClient.GetExternalWalletAsync("60af415a-61df-59d5-de2b-580cebc097fc", CancellationToken.None);
+        response.Result.Should().NotBeNull();
+    }
+
+    [Fact]
+    public async Task CreateExternalWalletAsync_should_add_a_new_wallet_in_fireblocks_database()
+    {
+        var walletName = _mockCreator.GetRandomString(10);
+        var refId = _mockCreator.GetRandomString(10);
+        var response = await _externalWalletsClient.CreateExternalWalletAsync(new Body12
         {
-            _externalWalletsClient = ServiceProvider.GetRequiredService<IExternalWalletsClient>();
-            _mockCreator = new MockCreator(output);
-        }
+            Name = walletName,
+            CustomerRefId = refId
+        });
+        var walletId = response.Result.Id;
 
-        [Fact]
-        public async Task GetExternalWalletsAsync_should_return_all_wallets()
-        {
-            var response = await _externalWalletsClient.GetExternalWalletsAsync();
-            response.Result.Should().NotBeEmpty();
-        }
+        // After write op, fireblocks needs some time to refresh data :(
+        await Task.Delay(2000).ConfigureAwait(false);
 
-        [Fact]
-        public async Task GetExternalWalletAsync_should_return_one_particular_wallet()
-        {
-            var response = await _externalWalletsClient.GetExternalWalletAsync("60af415a-61df-59d5-de2b-580cebc097fc", CancellationToken.None);
-            response.Result.Should().NotBeNull();
-        }
+        var response2 = await _externalWalletsClient.GetExternalWalletAsync(walletId);
+        response2.Result.Name.Should().Be(walletName);
+        await _externalWalletsClient.DeleteExternalWalletByIdAsync(walletId);
 
-        [Fact]
-        public async Task CreateExternalWalletAsync_should_add_a_new_wallet_in_fireblocks_database()
-        {
-            var walletName = _mockCreator.GetRandomString(10);
-            var refId = _mockCreator.GetRandomString(10);
-            var response = await _externalWalletsClient.CreateExternalWalletAsync(new Body12
-            {
-                Name = walletName,
-                CustomerRefId = refId
-            });
-            var walletId = response.Result.Id;
+        // After write op, fireblocks needs some time to refresh data :(
+        await Task.Delay(2000).ConfigureAwait(false);
 
-            // After write op, fireblocks needs some time to refresh data :(
-            await Task.Delay(2000).ConfigureAwait(false);
-
-            var response2 = await _externalWalletsClient.GetExternalWalletAsync(walletId);
-            response2.Result.Name.Should().Be(walletName);
-            await _externalWalletsClient.DeleteExternalWalletByIdAsync(walletId);
-
-            // After write op, fireblocks needs some time to refresh data :(
-            await Task.Delay(2000).ConfigureAwait(false);
-
-            await new Func<Task>(async () => await _externalWalletsClient.GetExternalWalletAsync(walletId)).Should()
-                .ThrowAsync<ApiException>();
-        }
+        await new Func<Task>(async () => await _externalWalletsClient.GetExternalWalletAsync(walletId)).Should()
+            .ThrowAsync<ApiException>();
     }
 }
