@@ -3,23 +3,22 @@ using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Trakx.Fireblocks.ApiClient.Tests.Unit
+namespace Trakx.Fireblocks.ApiClient.Tests.Unit;
+
+public class AddFireblocksClientExtensionTests : CredentialsTestsBase
 {
-    public class AddFireblocksClientExtensionTests : CredentialsTestsBase
+    public AddFireblocksClientExtensionTests(RsaKeyFixture fixture, ITestOutputHelper output) 
+        : base(fixture, output)
+    { }
+
+    [Fact]
+    public void Services_should_be_built()
     {
-        public AddFireblocksClientExtensionTests(RsaKeyFixture fixture, ITestOutputHelper output) 
-            : base(fixture, output)
-        { }
+        var serviceCollection = new ServiceCollection();
+        serviceCollection.AddFireblocksClient(Configuration);
 
-        [Fact]
-        public void Services_should_be_built()
-        {
-            var serviceCollection = new ServiceCollection();
-            serviceCollection.AddFireblocksClient(Configuration);
-
-            var serviceProvider = serviceCollection.BuildServiceProvider();
-            var client = serviceProvider.GetRequiredService<IExchangeAccountsClient>();
-            client.Should().NotBeNull();
-        }
+        var serviceProvider = serviceCollection.BuildServiceProvider();
+        var client = serviceProvider.GetRequiredService<IExchangeAccountsClient>();
+        client.Should().NotBeNull();
     }
 }
