@@ -9,56 +9,39 @@ using Trakx.Utils.Apis;
 using Trakx.Utils.DateTimeHelpers;
 using Trakx.Fireblocks.ApiClient.Utils;
 
-namespace Trakx.Fireblocks.ApiClient
+namespace Trakx.Fireblocks.ApiClient;
+
+public static partial class AddFireblocksClientExtension
 {
-    public static partial class AddFireblocksClientExtension
+    public static IServiceCollection AddFireblocksClient(
+        this IServiceCollection services, IConfiguration configuration)
     {
-        public static IServiceCollection AddFireblocksClient(
-            this IServiceCollection services, IConfiguration configuration)
-        {
-            services.AddOptions();
-            services.Configure<FireblocksApiConfiguration>(
-                configuration.GetSection(nameof(FireblocksApiConfiguration)));
-            AddCommonDependencies(services);
+        services.AddOptions();
+        services.Configure<FireblocksApiConfiguration>(
+            configuration.GetSection(nameof(FireblocksApiConfiguration)));
+        AddCommonDependencies(services);
 
-            return services;
-        }
+        return services;
+    }
 
-        public static IServiceCollection AddFireblocksClient(
-            this IServiceCollection services, FireblocksApiConfiguration apiConfiguration)
-        {
-            var options = Options.Create(apiConfiguration);
-            services.AddSingleton(options);
+    public static IServiceCollection AddFireblocksClient(
+        this IServiceCollection services, FireblocksApiConfiguration apiConfiguration)
+    {
+        var options = Options.Create(apiConfiguration);
+        services.AddSingleton(options);
 
-            AddCommonDependencies(services);
+        AddCommonDependencies(services);
 
-            return services;
-        }
+        return services;
+    }
 
-        private static void AddCommonDependencies(IServiceCollection services)
-        {
-            services.AddSingleton<IDateTimeProvider, DateTimeProvider>();
-            services.AddSingleton<IBearerCredentialsProvider, BearerCredentialsProvider>();
-            services.AddSingleton<IFireblocksCredentialsProvider, ApiKeyCredentialsProvider>();
+    private static void AddCommonDependencies(IServiceCollection services)
+    {
+        services.AddSingleton<IDateTimeProvider, DateTimeProvider>();
+        services.AddSingleton<IBearerCredentialsProvider, BearerCredentialsProvider>();
+        services.AddSingleton<IFireblocksCredentialsProvider, ApiKeyCredentialsProvider>();
 
-            services.AddSingleton<ClientConfigurator>();
-            AddClients(services);
-        }
-
-        private static void LogFailure(ILogger logger, DelegateResult<HttpResponseMessage> result, TimeSpan timeSpan, int retryCount, Context context)
-        {
-            if (result.Exception != null)
-            {
-                logger.Warning(result.Exception, "An exception occurred on retry {RetryAttempt} for {PolicyKey}. Retrying in {SleepDuration}ms.",
-                    retryCount, context.PolicyKey, timeSpan.TotalMilliseconds);
-            }
-            else
-            {
-                logger.Warning("A non success code {StatusCode} with reason {Reason} and content {Content} was received on retry {RetryAttempt} for {PolicyKey}. Retrying in {SleepDuration}ms.",
-                    (int)result.Result.StatusCode, result.Result.ReasonPhrase,
-                    result.Result.Content?.ReadAsStringAsync().GetAwaiter().GetResult(),
-                    retryCount, context.PolicyKey, timeSpan.TotalMilliseconds);
-            }
-        }
+        services.AddSingleton<ClientConfigurator>();
+        AddClients(services);
     }
 }
