@@ -5,17 +5,18 @@ using Polly;
 using Polly.Contrib.WaitAndRetry;
 using Polly.Extensions.Http;
 using Serilog;
+using Trakx.Utils.Apis;
 
-namespace Trakx.Fireblocks.ApiClient
-{
+namespace Trakx.Fireblocks.ApiClient;
+
     public static partial class AddFireblocksClientExtension
     {
         private static void AddClients(this IServiceCollection services)
         {
             var delay = Backoff.DecorrelatedJitterBackoffV2(medianFirstRetryDelay: TimeSpan.FromMilliseconds(100), retryCount: 2, fastFirst: true);
-                                    
+            
             services.AddHttpClient<IVaultClient, VaultClient>("Trakx.Fireblocks.ApiClient.VaultClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -24,13 +25,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<VaultClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("VaultClient"));
 
-                                
+        
             services.AddHttpClient<IInternalWalletsClient, InternalWalletsClient>("Trakx.Fireblocks.ApiClient.InternalWalletsClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -39,13 +40,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<InternalWalletsClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("InternalWalletsClient"));
 
-                                
+        
             services.AddHttpClient<IExternalWalletsClient, ExternalWalletsClient>("Trakx.Fireblocks.ApiClient.ExternalWalletsClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -54,13 +55,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<ExternalWalletsClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("ExternalWalletsClient"));
 
-                                
+        
             services.AddHttpClient<IExchangeAccountsClient, ExchangeAccountsClient>("Trakx.Fireblocks.ApiClient.ExchangeAccountsClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -69,13 +70,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<ExchangeAccountsClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("ExchangeAccountsClient"));
 
-                                
+        
             services.AddHttpClient<IFiatAccountsClient, FiatAccountsClient>("Trakx.Fireblocks.ApiClient.FiatAccountsClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -84,13 +85,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<FiatAccountsClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("FiatAccountsClient"));
 
-                                
+        
             services.AddHttpClient<ITransactionsClient, TransactionsClient>("Trakx.Fireblocks.ApiClient.TransactionsClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -99,13 +100,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<TransactionsClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("TransactionsClient"));
 
-                                
+        
             services.AddHttpClient<ISupportedAssetsClient, SupportedAssetsClient>("Trakx.Fireblocks.ApiClient.SupportedAssetsClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -114,13 +115,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<SupportedAssetsClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("SupportedAssetsClient"));
 
-                                
+        
             services.AddHttpClient<INetworkConnectionsClient, NetworkConnectionsClient>("Trakx.Fireblocks.ApiClient.NetworkConnectionsClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -129,13 +130,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<NetworkConnectionsClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("NetworkConnectionsClient"));
 
-                                
+        
             services.AddHttpClient<ITransferTicketsClient, TransferTicketsClient>("Trakx.Fireblocks.ApiClient.TransferTicketsClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -144,13 +145,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<TransferTicketsClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("TransferTicketsClient"));
 
-                                
+        
             services.AddHttpClient<IFeeClient, FeeClient>("Trakx.Fireblocks.ApiClient.FeeClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -159,12 +160,13 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<FeeClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("FeeClient"));
 
+        
             services.AddHttpClient<IWebhooksClient, WebhooksClient>("Trakx.Fireblocks.ApiClient.WebhooksClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
                     .Handle<ApiException>()
                     .Or<HttpRequestException>()
@@ -173,38 +175,40 @@ namespace Trakx.Fireblocks.ApiClient
                         onRetry: (result, timeSpan, retryCount, context) =>
                         {
                             var logger = Log.Logger.ForContext<WebhooksClient>();
-                            LogFailure(logger, result, timeSpan, retryCount, context);
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
                         })
                     .WithPolicyKey("WebhooksClient"));
 
+        
             services.AddHttpClient<IUsersClient, UsersClient>("Trakx.Fireblocks.ApiClient.UsersClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
-                        .Handle<ApiException>()
-                        .Or<HttpRequestException>()
-                        .OrTransientHttpStatusCode()
-                        .WaitAndRetryAsync(delay,
-                            onRetry: (result, timeSpan, retryCount, context) =>
-                            {
-                                var logger = Log.Logger.ForContext<UsersClient>();
-                                LogFailure(logger, result, timeSpan, retryCount, context);
-                            })
-                        .WithPolicyKey("UsersClient"));
+                    .Handle<ApiException>()
+                    .Or<HttpRequestException>()
+                    .OrTransientHttpStatusCode()
+                    .WaitAndRetryAsync(delay,
+                        onRetry: (result, timeSpan, retryCount, context) =>
+                        {
+                            var logger = Log.Logger.ForContext<UsersClient>();
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
+                        })
+                    .WithPolicyKey("UsersClient"));
 
+        
             services.AddHttpClient<IOffExchangeAccountsClient, OffExchangeAccountsClient>("Trakx.Fireblocks.ApiClient.OffExchangeAccountsClient")
-                .AddPolicyHandler((s, request) => 
+                .AddPolicyHandler((s, request) =>
                     Policy<HttpResponseMessage>
-                        .Handle<ApiException>()
-                        .Or<HttpRequestException>()
-                        .OrTransientHttpStatusCode()
-                        .WaitAndRetryAsync(delay,
-                            onRetry: (result, timeSpan, retryCount, context) =>
-                            {
-                                var logger = Log.Logger.ForContext<OffExchangeAccountsClient>();
-                                LogFailure(logger, result, timeSpan, retryCount, context);
-                            })
-                        .WithPolicyKey("OffExchangeAccountsClient"));
+                    .Handle<ApiException>()
+                    .Or<HttpRequestException>()
+                    .OrTransientHttpStatusCode()
+                    .WaitAndRetryAsync(delay,
+                        onRetry: (result, timeSpan, retryCount, context) =>
+                        {
+                            var logger = Log.Logger.ForContext<OffExchangeAccountsClient>();
+                            logger.LogApiFailure(result, timeSpan, retryCount, context);
+                        })
+                    .WithPolicyKey("OffExchangeAccountsClient"));
 
-        }
+        
     }
 }
