@@ -1,18 +1,16 @@
 ﻿using Trakx.Utils.Attributes;
 
-namespace Trakx.Fireblocks.ApiClient
-{
-    public record FireblocksApiConfiguration
-    {
-        
-        public string BaseUrl { get; set; }
-        
-#nullable disable
-        [SecretEnvironmentVariable]
-        public string ApiPubKey { get; init; }
+namespace Trakx.Fireblocks.ApiClient;
 
-        [SecretEnvironmentVariable]
-        public string ApiPrivateKey { get; init; }
+public record FireblocksApiConfiguration
+{
+    public string BaseUrl { get; set; }
+
+#nullable disable
+    [AwsParameter, SecretEnvironmentVariable]
+    public string ApiPubKey { get; init; }
+
+    [AwsParameter, SecretEnvironmentVariable]
+    public string ApiPrivateKey { get; init; }
 #nullable restore
-    }
 }
