@@ -1,97 +1,95 @@
 ﻿#pragma warning disable S1172, IDE0060 // Unused method parameters should be removed
 
-namespace Trakx.Fireblocks.ApiClient
-{
-    internal partial class VaultClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class InternalWalletsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class ExternalWalletsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class ExchangeAccountsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class FiatAccountsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class TransactionsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class SupportedAssetsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class NetworkConnectionsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class TransferTicketsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class FeeClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class WebhooksClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class UsersClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
-    internal partial class OffExchangeAccountsClient
-    {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
-        {
-            CredentialProvider.AddCredentials(request);
-        }
-    }
+namespace Trakx.Fireblocks.ApiClient;
 
+internal partial class VaultClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class InternalWalletsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class ExternalWalletsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class ExchangeAccountsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class FiatAccountsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class TransactionsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class SupportedAssetsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class NetworkConnectionsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class TransferTicketsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class FeeClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class WebhooksClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class UsersClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
+}
+internal partial class OffExchangeAccountsClient
+{
+    partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+    {
+        CredentialProvider.AddCredentials(request);
+    }
 }
